requirements.txt
torch>=1.12.0
torchvision>=0.13.0
numpy>=1.21.0
matplotlib>=3.5.0
imageio>=2.19.0
tqdm>=4.64.0
opencv-python>=4.6.0
scikit-image>=0.19.0
Pillow>=9.0.0
tkinter